feat: reorganize WatchlistAddInput component; move to AddToWatchListButton directory and remove unused WatchlistMenu
