+++ src/components/SymbolTable/AddToWatchListButton/WatchListMenu.tsx
@@ -3,7 +3,7 @@ import { ListTitle } from './WatchListMenu.style'
 import DeleteIcon from '@mui/icons-material/Delete'
 import StarRateRoundedIcon from '@mui/icons-material/StarRateRounded'
 import StarOutlineRoundedIcon from '@mui/icons-material/StarOutlineRounded'
-import { WatchlistAddInput } from '../Watchlist/WatchlistMenu/WatchlistAddInput'
+import { WatchlistAddInput } from './WatchlistAddInput'
 import { IWatchlist, useWatchlistStoreActions, useWatchlistStoreWatchlists } from '../../../stores/watchlistStore'
 import { useCallback } from 'react'
 import { ISymbolItem } from '../../../stores/symbataStore.types'
